refactor: change repo structure by adding core module
