+++ examples/prepare_result.py
@@ -9,8 +9,8 @@ from tqdm.contrib.concurrent import process_map
 from examples.config import MAX_WORKERS
 from examples.mono_test_generator import Clicker
 from examples.utils import SingleSolverResult, TestResult
-from mpest.mixture_distribution import DistributionInMixture, MixtureDistribution
-from mpest.types import Samples
+from mpest.annotations import Samples
+from mpest.core.mixture_distribution import DistributionInMixture, MixtureDistribution
 
 
 def nll(samples: Samples, mixture: MixtureDistribution) -> float:
