+++ mpest/em/breakpointers/step_count_breakpointer.py
@@ -1,7 +1,7 @@
 """Module which contains EM breakpointer by step count"""
 
+from mpest.core.mixture_distribution import MixtureDistribution
 from mpest.em.breakpointers.unionable_breakpointer import AUnionableBreakpointer
-from mpest.mixture_distribution import MixtureDistribution
 
 
 class StepCountBreakpointer(AUnionableBreakpointer):
