+++ mpest/em/distribution_checkers/finite_checker.py
@@ -2,10 +2,10 @@
 
 import numpy as np
 
+from mpest.core.mixture_distribution import DistributionInMixture
 from mpest.em.distribution_checkers.unionable_distribution_checker import (
     AUnionableDistributionChecker,
 )
-from mpest.mixture_distribution import DistributionInMixture
 
 
 class FiniteChecker(AUnionableDistributionChecker):
