+++ mpest/optimizers/scipy_slsqp.py
@@ -4,8 +4,8 @@ from collections.abc import Callable
 
 from scipy.optimize import minimize
 
+from mpest.annotations import Params
 from mpest.optimizers.abstract_optimizer import AOptimizer
-from mpest.types import Params
 
 
 class ScipySLSQP(AOptimizer):
